# Conflicts:
#	ios/graphics/Shader/Metal/LineShader.metal
#	shared/src/map/layers/tiled/vector/symbol/Tiled2dMapVectorSymbolGroup.cpp
#	shared/src/map/layers/tiled/vector/symbol/Tiled2dMapVectorSymbolGroup.h

+++ ios/graphics/Shader/Metal/LineShader.metal
@@ -256,57 +256,57 @@ lineGroupFragmentShader(LineVertexOut in [[stage_in]],
     if (segmentType == 0 || capType == 1 || (segmentType == 2 && t < 0.0) || (segmentType == 1 && t > 1.0)) {
       d = min(length(in.lineA), length(in.lineA - in.lineB));
     } else if (capType == 2) {
-      const half dLen = t < 0.0 ? -t * lineLength : (t - 1.0) * lineLength;
-      const half3 intersectPt = t * half3(in.lineB);
-      const half dOrth = abs(length(half3(in.lineA) - intersectPt));
+      const float dLen = t < 0.0 ? -t * lineLength : (t - 1.0) * lineLength;
+      const float3 intersectPt = t * float3(in.lineB);
+      const float dOrth = abs(length(float3(in.lineA) - intersectPt));
       d = max(dLen, dOrth);
     } else {
       d = 0;
       discard_fragment();
     }
   } else {
-    const half3 intersectPt = t * half3(in.lineB);
-    d = abs(length(half3(in.lineA) - intersectPt));
+    const float3 intersectPt = t * float3(in.lineB);
+    d = abs(length(float3(in.lineA) - intersectPt));
   }
 
   if (d > in.width) {
     discard_fragment();
   }
 
-  const half opacity = style->opacity;
-  const half colorA = style->color.a;
-  const half colorAGap = style->gapColor.a;
+  const float opacity = style->opacity;
+  const float colorA = style->color.a;
+  const float colorAGap = style->gapColor.a;
 
-  half a = colorA * opacity;
-  const half aGap = colorAGap * opacity;
+  float a = colorA * opacity;
+  const float aGap = colorAGap * opacity;
 
   const int dottedLine = int(style->dotted);
 
   if (in.scaledBlur > 0 && t > 0.0 && t < 1.0) {
-    const half nonBlurRange = (in.width - in.scaledBlur);
+    const float nonBlurRange = (in.width - in.scaledBlur);
     if (d > nonBlurRange) {
       a *= clamp(1 - max(0.0, d - nonBlurRange) / (in.scaledBlur), 0.0, 1.0);
     }
   }
 
   if (dottedLine == 1) {
-    const half skew = style->dottedSkew;
+    const float skew = style->dottedSkew;
 
-    const half factorToT = (in.width * 2) / lineLength * skew;
-    const half dashOffset = (in.width - skew * in.width) / lineLength;
+    const float factorToT = (in.width * 2) / lineLength * skew;
+    const float dashOffset = (in.width - skew * in.width) / lineLength;
 
-    const half dashTotalDotted =  2.0 * factorToT;
-    const half offset = half(in.lengthPrefix) / lineLength;
-    const half startOffsetSegmentDotted = fmod(offset, dashTotalDotted);
-    const half pos = t + startOffsetSegmentDotted;
+    const float dashTotalDotted =  2.0 * factorToT;
+    const float offset = float(in.lengthPrefix) / lineLength;
+    const float startOffsetSegmentDotted = fmod(offset, dashTotalDotted);
+    const float pos = t + startOffsetSegmentDotted;
 
-    const half intraDashPosDotted = fmod(pos, dashTotalDotted);
+    const float intraDashPosDotted = fmod(pos, dashTotalDotted);
       if ((intraDashPosDotted > 1.0 * factorToT + dashOffset && intraDashPosDotted < dashTotalDotted - dashOffset) ||
                               (length(half2(min(abs(intraDashPosDotted - 0.5 * factorToT), 0.5 * factorToT + dashTotalDotted - intraDashPosDotted) / (0.5 * factorToT + dashOffset), d / in.width)) > 1.0)) {
           discard_fragment();
       }
   } else if(numDash > 0) {
-    const half factorToT = (in.width * 2) / lineLength;
+    const float factorToT = (in.width * 2) / lineLength;
     const float dashTotal = style->dashArray.w * factorToT;
     const float startOffsetSegment = fmod(in.lengthPrefix / lineLength, dashTotal);
 
@@ -328,7 +328,7 @@ lineGroupFragmentShader(LineVertexOut in [[stage_in]],
 
               half relG = (intraDashPos - dxt) / (dyt - dxt);
               if (relG < (style->dash_fade * 0.5)) {
-                  half wG = relG / (style->dash_fade * 0.5);
+                  float wG = relG / (style->dash_fade * 0.5);
                   return half4(half3(style->gapColor.rgb), 1.0) * aGap * wG + half4(half3(style->color.rgb), 1.0) * a * (1.0 - wG);
               }
               else if (1.0 - relG < (style->dash_fade * 0.5)) {
@@ -348,7 +348,7 @@ lineGroupFragmentShader(LineVertexOut in [[stage_in]],
                   return half4(half3(style->gapColor.rgb), 1.0) * aGap * wG + half4(half3(style->color.rgb), 1.0) * a * (1.0 - wG);
               }
               else if (1.0 - relG < style->dash_fade) {
-                  half wG = (1.0 - relG) / style->dash_fade;
+                  float wG = (1.0 - relG) / style->dash_fade;
                   return half4(half3(style->gapColor.rgb), 1.0) * aGap * wG + half4(half3(style->color.rgb), 1.0) * a * (1.0 - wG);
               }
               else {
